centralize arithmetic and comparison tests
some de-duplication
